mosaik-web/requirements.txt
arrow==0.7.0
decorator==4.0.9
docopt==0.6.2
flake8==2.5.4
mccabe==0.4.0
mosaik-api==2.2
networkx==1.11
pep8==1.7.0
pkginfo==1.2.1
pyflakes==1.0.0
python-dateutil==2.4.2
requests==2.9.1
requests-toolbelt==0.6.0
simpy==3.0.8
simpy.io==0.2.2
six==1.10.0
twine==1.6.5
wheel==0.29.0
